finished develop03!!! I love drugs yippeee!!!

+++ prove/Develop04/Player.cs
@@ -14,7 +14,7 @@ public class Player
     {
         get
         {
-            return levelProgress/nextLevelReq;
+            return (double)levelProgress/nextLevelReq;
         }
     }
